--- a/.travis.yml
+++ b/.travis.yml
@@ -1,4 +1,5 @@
 language: go
+sudo: false
 install:
   - go get github.com/bmizerany/assert
 script: