While the regex version reads easier, the performance difference is very noticeable even for these easy replacements.

```
master:
BenchmarkManyDifferentSensors	       1	1928148881 ns/op
BenchmarkOneBigTimer	       1	1845274355 ns/op
BenchmarkLotsOfTimers	       1	1902944310 ns/op
BenchmarkParseLine	  300000	      5568 ns/op

non_regex:
BenchmarkManyDifferentSensors	       1	1838807819 ns/op
BenchmarkOneBigTimer	       1	1769153610 ns/op
BenchmarkLotsOfTimers	       1	1853729588 ns/op
BenchmarkParseLine	 1000000	      1131 ns/op
```